+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Fix https://github.com/humhub/humhub/issues/4858 delete mention on chrome android broken
 - Fix https://github.com/humhub/humhub/issues/4775 emoji search broken on android
 - Enh: Added base key mappings
+- Enh: Improved emoji chooser positioning on small devices
 
 
 1.1.0 (February 4, 2021)