+++ docs/CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.1.1 (Unreleased)
+1.1.1 (February 18, 2021)
 -----------------------------------------
 - Fix #27: Removed foat attribute from image markup
 - Fix https://github.com/humhub/humhub/issues/4858 delete mention on chrome android broken